Move NovncStatus Component to CommandAndControl app
