+++ app/packs/src/apps/command_and_control/NovncStatus.js
